Backend/VidlY/requirements.txt
requests
fastapi[all]
vaderSentiment
psycopg2-binary
httpx